Move some files into server
